backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
pinecone-client==5.0.1
python-dotenv==1.0.1
httpx==0.28.1
beautifulsoup4==4.12.3
